--- src/main/java/com/agilethought/atsceapi/domain/NewUserResponse.java
@@ -1,9 +1,0 @@
-package com.agilethought.atsceapi.domain;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class NewUserResponse {
-	private String id;
-}
