+++ src/main/java/com/agilethought/atsceapi/dto/NewUserRequest.java
@@ -1,0 +1,14 @@
+package com.agilethought.atsceapi.dto;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class NewUserRequest {
+	private Integer type;
+	private String firstName;
+	private String lastName;
+	private String email;
+	private String password;
+	private Integer status;
+}
