+++ src/main/java/com/agilethought/atsceapi/dto/NewUserResponse.java
@@ -1,0 +1,9 @@
+package com.agilethought.atsceapi.dto;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class NewUserResponse {
+	private String id;
+}
